Add index.js and configure tab navigation
